Refactor Docker setup and workflows for SPAR Mapper Partner API
- Updated Dockerfile to replace the SPAR Mapper API with the Partner API, including changes to dependencies and environment variables.
- Modified GitHub Actions workflows to build and publish the Partner API instead of the Mapper API.
- Removed obsolete files and references related to the Mapper API, ensuring a clean transition to the Partner API.
- Updated repository links in project metadata to reflect the new structure.
