+++ .github/workflows/publish.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     defaults:
       run:
-        working-directory: openg2p-spar-mapper-api
+        working-directory: openg2p-spar-mapper-partner-api
     steps:
     - name: Set up Python
       uses: actions/setup-python@v4
