+++ openg2p-spar-mapper-core/README.md
@@ -1,11 +1,8 @@
 # OpenG2P SPAR Mapper Core
 
-[![Pre-commit Status](https://github.com/OpenG2P/openg2p-spar-mapper-api/actions/workflows/pre-commit.yml/badge.svg?branch=develop)](https://github.com/OpenG2P/openg2p-spar-mapper-api/actions/workflows/pre-commit.yml?query=branch%3Adevelop)
-[![Build Status](https://github.com/OpenG2P/openg2p-spar-mapper-api/actions/workflows/test.yml/badge.svg?branch=develop)](https://github.com/OpenG2P/openg2p-spar-mapper-api/actions/workflows/test.yml?query=branch%3Adevelop)
-[![codecov](https://codecov.io/gh/OpenG2P/openg2p-spar-mapper-api/branch/develop/graph/badge.svg)](https://codecov.io/gh/OpenG2P/openg2p-spar-mapper-api)
-[![openapi](https://img.shields.io/badge/open--API-swagger-brightgreen)](https://validator.swagger.io/?url=https://raw.githubusercontent.com/OpenG2P/openg2p-spar-mapper-api/develop/api-docs/generated/openapi.json)
-[![PyPI](https://img.shields.io/pypi/v/openg2p-spar-mapper-api?label=pypi%20package)](https://pypi.org/project/openg2p-spar-mapper-api)
-[![PyPI - Downloads](https://img.shields.io/pypi/dm/openg2p-spar-mapper-api)](https://pypi.org/project/openg2p-spar-mapper-api)
+[![Pre-commit Status](https://github.com/OpenG2P/openg2p-spar/actions/workflows/pre-commit.yml/badge.svg?branch=develop)](https://github.com/OpenG2P/openg2p-spar/actions/workflows/pre-commit.yml?query=branch%3Adevelop)
+[![Build Status](https://github.com/OpenG2P/openg2p-spar/actions/workflows/test.yml/badge.svg?branch=develop)](https://github.com/OpenG2P/openg2p-spar/actions/workflows/test.yml?query=branch%3Adevelop)
+[![codecov](https://codecov.io/gh/OpenG2P/openg2p-spar/branch/develop/graph/badge.svg)](https://codecov.io/gh/OpenG2P/openg2p-spar)
 
 ## Licenses
