+++ openg2p-spar-mapper-partner-api/CODE-OF-CONDUCT.md
