+++ openg2p-spar-mapper-partner-api/main.py
@@ -2,7 +2,7 @@
 
 # ruff: noqa: I001
 
-from openg2p_spar_mapper_api.app import Initializer
+from openg2p_spar_mapper_partner_api.app import Initializer
 from openg2p_fastapi_common.ping import PingInitializer
 
 initializer = Initializer()
