+++ openg2p-spar-mapper-partner-api/pyproject.toml
@@ -3,11 +3,11 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "openg2p-spar-mapper-api"
+name = "openg2p-spar-mapper-partner-api"
 authors = [
   { name="OpenG2P", email="info@openg2p.org" },
 ]
-description = "OpenG2P SPAR Mapper API"
+description = "OpenG2P SPAR Mapper Partner API"
 readme = "README.md"
 requires-python = ">=3.7"
 classifiers = [
@@ -28,8 +28,8 @@ dynamic = ["version"]
 [project.urls]
 Homepage = "https://openg2p.org"
 Documentation = "https://docs.openg2p.org/"
-Repository = "https://github.com/OpenG2P/openg2p-spar-mapper-api"
-Source = "https://github.com/OpenG2P/openg2p-spar-mapper-api"
+Repository = "https://github.com/OpenG2P/openg2p-spar"
+Source = "https://github.com/OpenG2P/openg2p-spar"
 
 [tool.hatch.version]
-path = "src/openg2p_spar_mapper_api/__init__.py"
+path = "src/openg2p_spar_mapper_partner_api/__init__.py"
