+++ openg2p-spar-mapper-partner-api/test_dfsp_endpoints.sh
